refator. reducer, actions 파일 분리
