+++ src/redux/actions.js
@@ -1,0 +1,23 @@
+export function updateTaskTitle(taskTitle) {
+  return {
+    type: 'updateTaskTitle',
+    payload: {
+      taskTitle,
+    },
+  };
+}
+
+export function addTask() {
+  return {
+    type: 'addTask',
+  };
+}
+
+export function deleteTask(id) {
+  return {
+    type: 'deleteTask',
+    payload: {
+      id,
+    },
+  };
+}
